Move SweeplineIntersector to its own file. Remove Queue, replace observed_segments with Set, condense event creation branching.
