+++ lib/rgeo/cartesian.rb
@@ -13,3 +13,4 @@ require_relative "cartesian/factory"
 require_relative "cartesian/interface"
 require_relative "cartesian/bounding_box"
 require_relative "cartesian/analysis"
+require_relative "cartesian/sweepline_intersector"
